[Perf] Don't delete output dir on bots
Modify run_performance_tests.py script script to skip deleting output
directory on test bots, as the Result Sink might still be uploading
files when this script ends.

Bug: 1104244
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2875151
Commit-Queue: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Auto-Submit: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#879450}

+++ testing/scripts/run_performance_tests.py
@@ -503,10 +503,14 @@ def execute_telemetry_benchmark(
            'outputing structured test results and perf results output:')
     print traceback.format_exc()
   finally:
-    # Add ignore_errors=True because otherwise rmtree may fail due to leaky
-    # processes of tests are still holding opened handles to files under
-    # |tempfile_dir|. For example, see crbug.com/865896
-    shutil.rmtree(temp_dir, ignore_errors=True)
+    # On swarming bots, don't remove output directory, since Result Sink might
+    # still be uploading files to Result DB. Also, swarming bots automatically
+    # clean up at the end of each task.
+    if 'SWARMING_TASK_ID' not in os.environ:
+      # Add ignore_errors=True because otherwise rmtree may fail due to leaky
+      # processes of tests are still holding opened handles to files under
+      # |tempfile_dir|. For example, see crbug.com/865896
+      shutil.rmtree(temp_dir, ignore_errors=True)
 
   print_duration('executing benchmark %s' % command_generator.benchmark, start)
